split logic into tabs components
