+++ src/pages/TabsPage.tsx
@@ -1,37 +1,8 @@
-import React from 'react';
-import { Link, useParams } from 'react-router-dom';
+import { Tabs } from '../components/Tabs';
 
-export const TabsPage = () => {
-  const tabs = [
-    { id: 'tab-1', title: 'Tab 1', content: 'Some text 1' },
-    { id: 'tab-2', title: 'Tab 2', content: 'Some text 2' },
-    { id: 'tab-3', title: 'Tab 3', content: 'Some text 3' },
-  ];
-
-  const { tabId } = useParams();
-  const selectedTab = tabs.find(tab => tab.id === tabId);
-
-  return (
-    <>
-      <h1 className="title">Tabs page</h1>
-
-      <div className="tabs is-boxed">
-        <ul>
-          {tabs.map(tab => (
-            <li
-              key={tab.id}
-              data-cy="Tab"
-              className={tab.id === tabId ? 'is-active' : ''}
-            >
-              <Link to={`/tabs/${tab.id}`}>{tab.title}</Link>
-            </li>
-          ))}
-        </ul>
-      </div>
-
-      <div className="block" data-cy="TabContent">
-        {selectedTab ? selectedTab.content : 'Please select a tab'}
-      </div>
-    </>
-  );
-};
+export const TabsPage = () => (
+  <>
+    <h1 className="title">Tabs page</h1>
+    <Tabs />
+  </>
+);
